+++ src/output/consul/plain_output.rs
@@ -27,8 +27,6 @@ impl Default for PlainOutputCatalogResult {
 
 impl OutputCatalogResult for PlainOutputCatalogResult {
     fn output<T: Write>(&self, writer: &mut T, catalog: &Catalog) -> Result<()> {
-        // We have to create / allocate the Strings first since `Table` only accepts `&str` and some
-        // `InstanceDescriptorFields` need to allocate representations first, e.g., `InstanceDescriptorFields::Tags`
         let mut rows = Vec::new();
         for service in catalog.services() {
             if let Some(nodes) = catalog.nodes_by_service(service) {