+++ src/output/instances/mod.rs
@@ -5,9 +5,11 @@ use output::*;
 use utils::command::CommandResult;
 
 pub mod json_output;
+pub mod plain_output;
 pub mod table_output;
 
 pub use self::json_output::{JsonOutputCommandResults, JsonOutputInstances, JsonOutputStateChanges};
+pub use self::plain_output::{PlainOutputInstances};
 pub use self::table_output::{TableOutputCommandResults, TableOutputInstances, TableOutputStatusChanges};
 
 pub trait OutputInstances {