+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixes #3: `ceres ops issues create` wait for with prompt in order to ensure that your editor finished unless `--no-wait` is set.
+- Fixes #25: "_Allow ceres instances ssh to receive Instance ID from stdin_" by adding plain output to `instances list`.
 
 
 ## [0.0.13] - 2018-05-17